Current version
This is the state the program is actually in now, added in a few more
classes to work with, removed dependency on .txt files for a single
sculpture, and started moving towards creating objects that will
eventually be able to move around other pieces.

Global values are added, and may eventually be moved to a config.yaml
file, but most are there for sake of coordinate placement, and color
values.

Cubie objects are the individual cubes that have sticker faces, and are
based around their own local coordinates. This way cubies can be moved
around, rotated, and the engine doesn't have to worry about moving
around specific points, and only has to worry about moving and rotating
objects.