Code runs again
First version that actually runs the code since overhauling to a new
system based around objects that contain polygons instead of based
around a master polygon system.

Unfortunately my new method for rendering the polygons doesn't seem to
work properly, as all polygons are rendered as being either very far
away or extremely small, when they should be very close to the camera.

Going to run through the math, and fix this issue so things show up as
large as they need to be.

Also note, this was the first test to see if my new rendering method
would work, and it doesn't seem to work as intended, so another reason
to check the math again.