Unexpected results from 3D renderer
Changed out the one object in the 3D renderer back to the static
sculpture, and added in something to change around the camera z
coordinate so to test the 3D renderer and attempt to find out why it's
acting up.

No clue yet on why it's acting up, but it produces some very interesting
results when zooming in between z=-1.7 to z=1.7-ish. Numpad 4/1 to move
forward and backwards by mult, and numpad 5/2 to increase/decrease mult
by a factor of 10. This way you can move in more quickly and more slowly
if you so choose.