Things working... kinda...
Added a few more things in for moving the camera around during the simulation, and got the renderer to work better, but still runs into quirky issues. Previous version, rubik's cube sculpture starts appearing very far away, then quickly gets very large as the camera approaches it on the Z-Axis. Remedied this by reworking the render method so the virtual screen it plots points to has a given width and height, then it finds a percentage down, over coordinate, and then converts that to a resolution down, over coordinate. This has fixed the issue of things initially appearing very small, because the difference of small degrees on the perspective angles reflects to be much larger of the screen's width and height rather than between 0-1 when far away, and 1-infinity when almost inside the cube.

This was the first priority issue with the previous rendering method, so things look correct from the right angle, but it now runs into issues when attempting to look at the sculptures from side angles.