Few minor fixes
Changed default camera position so the entire cube is visible from startup, added the polygon data back into the repo, and changed it's path in the program to the repo directory instead of the location of it on my local machine, fixed an issue of the renderer displaying things upside down, and of the screen-size not propagating through all the objects leading to resolution mismatching on different size screens. Commented out initialization of Cubie object while still trying to fix the renderer. Lastly, fixed typo at the try/except so when you close out of the pygame window, the python window disappears as well. I know except: without an error isn't preferred, but I like having it so I can at least debug issues with a traceback.print_exc() and time.sleep(), and I usually only ever use it around a main function or something similar.

Still working on trying to figure out what's wrong with the renderer that it doesn't rotate objects when looking around it. It seems to do something along the lines of move the observed object further to the side, and rotating the camera seems to only extend the view at the same perspective instead of applying a new perspective, which would explain why the view appears more and more warped the further to the side it is. Not sure how to fix this issue with the renderer, because I'm still not sure why this is happening.