Rendering shows up correctly! Most of the time!
Debugged an issue with the math behind the rendering method. The problem was while figuring out the camera's perspective of points, theta (left/right) angle was only concerned with changes in the X/Z plane, and phi (up/down) was only concerned with changes in the Y/Z plane. This was changed so both angles are now concerned with changes in all dimensions, as all dimensions affect the perspective. After that, added in 8/9/0 keys for testing perspective. 0 will rotate the camera around the Y axis, and 9 will rotate the camera around the X axis. 8 is used to reverse the increment and applies to both. May eventually test out rotation around the Z axis, but that's not a primary concern right now when the renderer is still running into issues around the other two axises as is.

Anyways, as it stands right now, the program seems to work just fine between angles +/- 0.9*pi, before it rapidly compresses the view of the object, then spinning it around the camera. Or at least that seems to be what's happening, hard to tell. Back to the drawing board.