Abandoning full functionality of rendering system to work on Twisty Puzzle emulation.
I've spent quite a bit of time trying to get the rendering system to work, and am running into too many issues with it to justify working on it much any longer. I may eventually return to it to finish it, but for now I'd rather get back to work on the other half of why I started this project, that being to emulate twisty puzzles of all types like the Rubik's Cube.

My primary end goal for this project is to emulate as well as possible as many types of twisty puzzles as I can, using a robust system that allows for many types of puzzles to be enjoyed in a virtual experience. Beyond that, I also hope to add in a solver that can also act as a tutor of sorts for people learning to solve different puzzles or learning new methods/algorithms. Lastly, I also want this program to be able to try to find algorithms for the user. This would be using a fairly simple setup of "my cube is in this situation, and I need it to be in this situation, and these are the cubes I don't care what color will be by the end of the algorithm." I intend to add a few more small features, but that's all I'm going to explain for now.

For the time being, I'm going to work on implementing some sort of cube emulation and interaction, as well as getting the program to move the cube according to a given move-set, and see what I can do from there. I'll still use the old rendering system, but the camera will be fixed in place until I can get it working properly when it's moving.