Setting up framework for supporting many puzzles.
Program is intended to support many different types of cubes, and I'm working on a system that should be able to handle most twisty puzzles. The Rubik's Cube folder is the start of such a system, where inside the Cube Designs folder, the program would have a separate folder for each cube that it can simulate. Inside that folder, it would look for an init file (deciding whether to support .yaml files or just swap all files to .dat) then load all such related data and files. The new 1.txt file is an example of the dict file that would be contained in such an init file.

From there, the init dict would begin loading cubie.dat and sticker.dat files as necessary and associating them with such related parts to a virtual cube. It would begin creating instances of different cubie designs (the individual pieces of a twisty puzzle) and laying stickers in relation to those cubies, establishing associations to specific cubies so they move in sync with each other.

Afterwards, the init file would contain the possible set of moves that the cube can perform. Using this data, it would be possible for the program to know what moves mean what, what to do when instructed to move specific pieces, and be able to perform moves on it's own, thereby allowing it to eventually perform algorithms and solve cubes on it's own.

That's as far as I've gotten for now, but I'm expecting to have a fair bit more progress after this yet, come next commit.